added health, shooting and projectile components
shooting uses angle for direction so sprite component needed to render with 0 angle because entities will use spritesheets,

added shooting system that handles shooting with Space key

+++ 2DGameEngine/src/Components/AnimationComponent.cpp
@@ -4,5 +4,5 @@
 AnimationComponent::AnimationComponent(const int numFrames, const int frameSpeedRate, const bool looping)
 	: FrameSpeedRate(frameSpeedRate), Looping(looping), NumFrames(numFrames)
 {
-	StartTime = SDL_GetTicks();
+	StartTime = static_cast<int>(SDL_GetTicks());
 }
